Update PacketCategorizingHandler to use console for output
The major change in this commit is the update of PacketCategorizingHandler to use an IAnsiConsole parameter for output. A new dictionary and a private method have been added to categorize and color-code the output based on TCP packet destination port. A dependency to the Spectre.Console package has been added and the AsciiChart.Sharp package has been removed. Additionally, a new implementation of IAnsiConsole, MyAnsiConsole, has been added to the project.